isolate crawling to new script, existing script just do processing
